requirements.txt
PyGithub
wget
argparse
